+++ @types/index.d.ts
@@ -157,6 +157,13 @@ declare module "opentok-react-native" {
      * Event handlers passed into the native session instance.
      */
     eventHandlers?: OTSessionEventHandlers;
+
+
+    /**
+     * Used to set the encryption secret in a session that uses end-to-end encryption.
+     * See https://tokbox.com/developer/guides/end-to-end-encryption.
+     */
+    encryptionSecret?: string
   }
 
   interface OTSessionSessionOptions {
@@ -237,7 +244,7 @@ declare module "opentok-react-native" {
     connectionDestroyed?: CallbackWithParam<ConnectionDestroyedEvent, any>;
 
     /**
-     * Sent if the attempt to connect to the session fails or if the connection to the session drops due to an error after a successful connection.
+     * Sent if the attempt to connect to the session fails or if the connection to the session drops due to an error after a successful connection. Also sent if setting an encryption secret results in an error.
      */
     error?: CallbackWithParam<ErrorEvent, any>;