+++ src/OTSession.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import { pick, isNull } from 'underscore';
 import { setNativeEvents, removeNativeEvents,  OT } from './OT';
 import { sanitizeSessionEvents, sanitizeSessionOptions, sanitizeSignalData,
-   sanitizeCredentials, getConnectionStatus } from './helpers/OTSessionHelper';
+   sanitizeCredentials, sanitizeEncryptionSecret, getConnectionStatus } from './helpers/OTSessionHelper';
 import { handleError } from './OTError';
 import { logOT, getOtrnErrorEventHandler } from './helpers/OTHelper';
 import OTContext from './contexts/OTContext';
@@ -29,9 +29,10 @@ export default class OTSession extends Component {
   }
   componentDidMount() {
     const sessionOptions = sanitizeSessionOptions(this.props.options);
+    const encryptionSecret = sanitizeEncryptionSecret(this.props.encryptionSecret);
     const { apiKey, sessionId, token } = this.sanitizedCredentials;
     if (apiKey && sessionId && token) {
-      this.createSession(this.sanitizedCredentials, sessionOptions);
+      this.createSession(this.sanitizedCredentials, sessionOptions, encryptionSecret);
       logOT({ apiKey, sessionId, action: 'rn_initialize', proxyUrl: sessionOptions.proxyUrl });
     } else {
       handleError('Please check your OpenTok credentials.');
@@ -48,19 +49,28 @@ export default class OTSession extends Component {
     const updateSessionProperty = (key, defaultValue) => {
       if (shouldUpdate(key, defaultValue)) {
         const value = useDefault(this.props[key], defaultValue);
-        this.signal(value);
+        if (key === 'signal') {
+          this.signal(value);
+        }
+        if (key === 'encryptionSecret') {
+          this.setEncryptionSecret(value);
+        }
       }
     };
 
     updateSessionProperty('signal', {});
+    updateSessionProperty('encryptionSecret', undefined);
   }
   componentWillUnmount() {
     this.disconnectSession();
   }
-  createSession(credentials, sessionOptions) {
+  createSession(credentials, sessionOptions, encryptionSecret) {
     const { signal } = this.props;
     const { apiKey, sessionId, token } = credentials;
     OT.initSession(apiKey, sessionId, sessionOptions);
+    if (encryptionSecret) {
+      this.setEncryptionSecret(encryptionSecret);
+    }
     OT.connect(sessionId, token, (error) => {
       if (error) {
         this.otrnEventHandler(error);
@@ -128,6 +138,14 @@ export default class OTSession extends Component {
   disableForceMute() {
     return OT.disableForceMute(this.props.sessionId);
   }
+  setEncryptionSecret(secret) {
+    const errorHandler = this.props.eventHandlers.error;
+    OT.setEncryptionSecret(this.props.sessionId, sanitizeEncryptionSecret(secret), (error) => {
+      if (error && errorHandler) {
+        errorHandler(error);
+      }
+    });
+  }
   render() {
     const { style, children, sessionId, apiKey, token } = this.props;
     const { sessionInfo } = this.state;
@@ -156,6 +174,7 @@ OTSession.propTypes = {
   eventHandlers: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   options: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   signal: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  encryptionSecret: PropTypes.string,
 };
 
 OTSession.defaultProps = {